+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Security
 
 ## [1.8.5] - 2023-08-05
+### Added
+- tests.py to settings
 ### Changed
 - __core__ directory to f1t